+++ internal/provider/vm_resource.go
@@ -3,14 +3,26 @@ package provider
 import (
 	"context"
 	"fmt"
+	"regexp"
 	"time"
 
 	"github.com/CudoVentures/terraform-provider-cudo/internal/compute/vm"
 	"github.com/CudoVentures/terraform-provider-cudo/internal/helper"
 
 	grpc_retry "github.com/grpc-ecosystem/go-grpc-middleware/retry"
+	"github.com/hashicorp/terraform-plugin-framework-validators/int64validator"
+	"github.com/hashicorp/terraform-plugin-framework-validators/stringvalidator"
 	"github.com/hashicorp/terraform-plugin-framework/path"
 	"github.com/hashicorp/terraform-plugin-framework/resource"
+	"github.com/hashicorp/terraform-plugin-framework/resource/schema"
+	"github.com/hashicorp/terraform-plugin-framework/resource/schema/boolplanmodifier"
+	"github.com/hashicorp/terraform-plugin-framework/resource/schema/int64default"
+	"github.com/hashicorp/terraform-plugin-framework/resource/schema/int64planmodifier"
+	"github.com/hashicorp/terraform-plugin-framework/resource/schema/listplanmodifier"
+	"github.com/hashicorp/terraform-plugin-framework/resource/schema/planmodifier"
+	"github.com/hashicorp/terraform-plugin-framework/resource/schema/setplanmodifier"
+	"github.com/hashicorp/terraform-plugin-framework/resource/schema/stringplanmodifier"
+	"github.com/hashicorp/terraform-plugin-framework/schema/validator"
 	"github.com/hashicorp/terraform-plugin-framework/types"
 	"github.com/hashicorp/terraform-plugin-log/tflog"
 	"google.golang.org/genproto/googleapis/type/decimal"
@@ -27,6 +39,208 @@ func NewVMResource() resource.Resource {
 
 func (r *VMResource) Metadata(ctx context.Context, req resource.MetadataRequest, resp *resource.MetadataResponse) {
 	resp.TypeName = "cudo_vm"
+}
+
+func (r *VMResource) Schema(ctx context.Context, req resource.SchemaRequest, resp *resource.SchemaResponse) {
+	resp.Schema = schema.Schema{
+		// This description is used by the documentation generator and the language server.
+		MarkdownDescription: "VM resource",
+		Attributes: map[string]schema.Attribute{
+			"boot_disk": schema.SingleNestedAttribute{
+				MarkdownDescription: "Specification for boot disk",
+				Attributes: map[string]schema.Attribute{
+					"size_gib": schema.Int64Attribute{
+						PlanModifiers: []planmodifier.Int64{
+							int64planmodifier.RequiresReplace(),
+						},
+						Computed:            true,
+						Optional:            true,
+						MarkdownDescription: "Size of boot disk in Gib",
+					},
+					"image_id": schema.StringAttribute{
+						PlanModifiers: []planmodifier.String{
+							stringplanmodifier.RequiresReplace(),
+						},
+						MarkdownDescription: "ID of OS image on boot disk",
+						Required:            true,
+						Validators:          []validator.String{stringvalidator.RegexMatches(regexp.MustCompile("^[a-z]([a-z0-9-]{0,61}[a-z0-9])?$"), "must be a valid resource id")},
+					},
+				},
+				Required: true,
+			},
+			"cpu_model": schema.StringAttribute{
+				PlanModifiers: []planmodifier.String{
+					stringplanmodifier.RequiresReplace(),
+				},
+				MarkdownDescription: "The model of the CPU.",
+				Optional:            true,
+				Computed:            true,
+			},
+			"data_center_id": schema.StringAttribute{
+				PlanModifiers: []planmodifier.String{
+					stringplanmodifier.RequiresReplace(),
+				},
+				MarkdownDescription: "The id of the datacenter where the VM instance is located.",
+				Optional:            true,
+				Computed:            true,
+				Validators:          []validator.String{stringvalidator.RegexMatches(regexp.MustCompile("^[a-z]([a-z0-9-]{0,61}[a-z0-9])?$"), "must be a valid resource id")},
+			},
+			"external_ip_address": schema.StringAttribute{
+				MarkdownDescription: "The external IP address of the VM instance.",
+				Computed:            true,
+			},
+			"gpu_model": schema.StringAttribute{
+				PlanModifiers: []planmodifier.String{
+					stringplanmodifier.RequiresReplace(),
+				},
+				MarkdownDescription: "The model of the GPU.",
+				Optional:            true,
+				Computed:            true,
+			},
+			"gpus": schema.Int64Attribute{
+				PlanModifiers: []planmodifier.Int64{
+					int64planmodifier.RequiresReplace(),
+				},
+				MarkdownDescription: "Number of GPUs",
+				Optional:            true,
+				Computed:            true,
+				Default:             int64default.StaticInt64(0),
+			},
+			"id": schema.StringAttribute{
+				PlanModifiers: []planmodifier.String{
+					stringplanmodifier.RequiresReplace(),
+				},
+				MarkdownDescription: "ID for VM within project",
+				Required:            true,
+				Validators:          []validator.String{stringvalidator.RegexMatches(regexp.MustCompile("^[a-z]([a-z0-9-]{0,61}[a-z0-9])?$"), "must be a valid resource id e.g. my-vm")},
+			},
+			"internal_ip_address": schema.StringAttribute{
+				MarkdownDescription: "The internal IP address of the VM instance.",
+				Computed:            true,
+			},
+			"machine_type": schema.StringAttribute{
+				PlanModifiers: []planmodifier.String{
+					stringplanmodifier.RequiresReplace(),
+				},
+				MarkdownDescription: "VM machine type, from machine type data source",
+				Optional:            true,
+				Computed:            true,
+				Validators:          []validator.String{stringvalidator.RegexMatches(regexp.MustCompile("^[a-z]([a-z0-9-]{0,61}[a-z0-9])?$"), "must be a valid resource id")},
+			},
+			"max_price_hr": schema.StringAttribute{
+				PlanModifiers: []planmodifier.String{
+					stringplanmodifier.RequiresReplace(),
+				},
+				MarkdownDescription: "The maximum price per hour for the VM instance.",
+				Optional:            true,
+			},
+			"memory_gib": schema.Int64Attribute{
+				PlanModifiers: []planmodifier.Int64{
+					int64planmodifier.RequiresReplace(),
+				},
+				MarkdownDescription: "Amount of VM memory in GiB",
+				Optional:            true,
+			},
+			"networks": schema.ListNestedAttribute{
+				Optional:            true,
+				MarkdownDescription: "Network adapters for private networks",
+				NestedObject: schema.NestedAttributeObject{
+					Attributes: map[string]schema.Attribute{
+						"network_id": schema.StringAttribute{
+							PlanModifiers: []planmodifier.String{
+								stringplanmodifier.RequiresReplace(),
+							},
+							MarkdownDescription: "ID of private network to attach the NIC to",
+							Required:            true,
+						},
+						"assign_public_ip": schema.BoolAttribute{
+							PlanModifiers: []planmodifier.Bool{
+								boolplanmodifier.RequiresReplace(),
+							},
+							MarkdownDescription: "Assign a public IP to the NIC",
+							Optional:            true,
+						},
+						"external_ip_address": schema.StringAttribute{
+							MarkdownDescription: "The external IP address of the NIC.",
+							Computed:            true,
+						},
+						"internal_ip_address": schema.StringAttribute{
+							MarkdownDescription: "The internal IP address of the NIC.",
+							Computed:            true,
+						},
+						"security_group_ids": schema.SetAttribute{
+							ElementType:         types.StringType,
+							Optional:            true,
+							MarkdownDescription: "Security groups to assign to the NIC",
+						},
+					},
+				},
+			},
+			"password": schema.StringAttribute{
+				PlanModifiers: []planmodifier.String{
+					stringplanmodifier.RequiresReplace(),
+				},
+				MarkdownDescription: "Root password for linux, or Admin password for windows",
+				Optional:            true,
+				Sensitive:           true,
+				Validators:          []validator.String{stringvalidator.LengthBetween(6, 64)},
+			},
+			"price_hr": schema.StringAttribute{
+				MarkdownDescription: "The current price per hour for the VM instance.",
+				Computed:            true,
+			},
+			"project_id": schema.StringAttribute{
+				PlanModifiers: []planmodifier.String{
+					stringplanmodifier.RequiresReplace(),
+				},
+				MarkdownDescription: "The project the VM instance is in.",
+				Optional:            true,
+			},
+			"renewable_energy": schema.BoolAttribute{
+				MarkdownDescription: "Whether the VM instance is powered by renewable energy",
+				Computed:            true,
+			},
+			"security_group_ids": schema.SetAttribute{
+				PlanModifiers: []planmodifier.Set{
+					setplanmodifier.RequiresReplace(),
+				},
+				ElementType:         types.StringType,
+				Optional:            true,
+				MarkdownDescription: "Security groups to assign to the VM when using public networking",
+			},
+			"ssh_key_source": schema.StringAttribute{
+				PlanModifiers: []planmodifier.String{
+					stringplanmodifier.RequiresReplace(),
+				},
+				MarkdownDescription: "Which SSH keys to add to the VM: project (default), user or custom",
+				Optional:            true,
+				Validators:          []validator.String{stringvalidator.OneOf("project", "user", "custom")},
+			},
+			"ssh_keys": schema.ListAttribute{
+				PlanModifiers: []planmodifier.List{
+					listplanmodifier.RequiresReplace(),
+				},
+				ElementType:         types.StringType,
+				MarkdownDescription: "List of SSH keys to add to the VM, ssh_key_source must be set to custom",
+				Optional:            true,
+			},
+			"start_script": schema.StringAttribute{
+				PlanModifiers: []planmodifier.String{
+					stringplanmodifier.RequiresReplace(),
+				},
+				MarkdownDescription: "A script to run when VM boots",
+				Optional:            true,
+			},
+			"vcpus": schema.Int64Attribute{
+				PlanModifiers: []planmodifier.Int64{
+					int64planmodifier.RequiresReplace(),
+				},
+				MarkdownDescription: "Number of VCPUs",
+				Optional:            true,
+				Validators:          []validator.Int64{int64validator.AtMost(100)},
+			},
+		},
+	}
 }
 
 // VMResource defines the resource implementation.
@@ -196,8 +410,7 @@ func (r *VMResource) Create(ctx context.Context, req resource.CreateRequest, res
 		return
 	}
 
-	fillVmState(state, vm.VM)
-	resp.Diagnostics.Append(resp.State.Set(ctx, &state)...)
+	resp.Diagnostics.Append(resp.State.Set(ctx, fillVmState(vm.VM))...)
 }
 
 func (r *VMResource) Read(ctx context.Context, req resource.ReadRequest, resp *resource.ReadResponse) {
@@ -228,8 +441,7 @@ func (r *VMResource) Read(ctx context.Context, req resource.ReadRequest, resp *r
 		return
 	}
 
-	fillVmState(state, vm.VM)
-	resp.Diagnostics.Append(resp.State.Set(ctx, &state)...)
+	resp.Diagnostics.Append(resp.State.Set(ctx, fillVmState(vm.VM))...)
 }
 
 func (r *VMResource) Update(ctx context.Context, req resource.UpdateRequest, resp *resource.UpdateResponse) {
@@ -257,11 +469,10 @@ func (r *VMResource) Delete(ctx context.Context, req resource.DeleteRequest, res
 	var state *VMResourceModel
 	resp.Diagnostics.Append(req.State.Get(ctx, &state)...)
 
-	params := &vm.TerminateVMRequest{}
-	params.ProjectId = r.client.DefaultProjectID
-	params.Id = state.ID.ValueString()
+	projectId := r.client.DefaultProjectID
+	vmId := state.ID.ValueString()
 
-	if _, err := waitForVmAvailable(ctx, params.ProjectId, params.Id, r.client.VMClient); err != nil {
+	if _, err := waitForVmAvailable(ctx, projectId, vmId, r.client.VMClient); err != nil {
 		resp.Diagnostics.AddError(
 			"Unable to wait for VM resource to be available",
 			err.Error(),
@@ -269,7 +480,10 @@ func (r *VMResource) Delete(ctx context.Context, req resource.DeleteRequest, res
 		return
 	}
 
-	_, err := r.client.VMClient.TerminateVM(ctx, params)
+	_, err := r.client.VMClient.TerminateVM(ctx, &vm.TerminateVMRequest{
+		ProjectId: projectId,
+		Id:        vmId,
+	})
 	if err != nil {
 		resp.Diagnostics.AddError(
 			"Unable to delete VM resource",
@@ -278,7 +492,7 @@ func (r *VMResource) Delete(ctx context.Context, req resource.DeleteRequest, res
 		return
 	}
 
-	_, err = waitForVmDelete(ctx, params.ProjectId, params.Id, r.client.VMClient)
+	_, err = waitForVmDelete(ctx, projectId, vmId, r.client.VMClient)
 	if err != nil {
 		resp.Diagnostics.AddError(
 			"Unable to wait for VM resource to be deleted",
@@ -379,7 +593,8 @@ func waitForVmDelete(ctx context.Context, projectId string, vmID string, c vm.VM
 	}
 }
 
-func fillVmState(state *VMResourceModel, vm *vm.VM) {
+func fillVmState(vm *vm.VM) VMResourceModel {
+	var state VMResourceModel
 	state.DataCenterID = types.StringValue(vm.DatacenterId)
 	state.CPUModel = types.StringValue(vm.CpuModel)
 	state.GPUs = types.Int64Value(int64(vm.GpuQuantity))
@@ -401,4 +616,5 @@ func fillVmState(state *VMResourceModel, vm *vm.VM) {
 	state.ExternalIPAddress = types.StringValue(vm.ExternalIpAddress)
 	state.PriceHr = types.StringValue(fmt.Sprintf("%0.2f", vm.PriceHr))
 	state.RenewableEnergy = types.BoolValue(vm.RenewableEnergy)
+	return state
 }
